+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Release Notes
 
-## WIP
+## 2.3.3 (2022-11-02)
 
 - Add ESC + auto focus support for dialogs
 - Update Vue to 3.2.41
 - Add finder module (only for super admin users)
 - Set default value to false for boolean fields
+- Update vendor libs
 
 ## 2.3.2 (2022-10-24)